ragas_demo/requirements.txt
ragas>=0.1.0
datasets>=2.14.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0
pandas>=2.0.0
openai>=1.0.0

